+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changes
 * [UI]: Fixed bug preventing associated projects from being loaded into the project samples table. (19.01.3)
 * [Developer]: Moved FastQC results out of database to filesystem for a big reduction in database size and performance.  See https://irida.corefacility.ca/documentation/adminstrator/upgrades/#1904 for more information. 
 * [UI]: Added hard wrap on sample name on sample details page.
+* [UI]: New dedicated cart page.
+* [UI]: Exporting to Galaxy now runs through the new cart interface.  Export to Galaxy through the Project Samples page has been deprecated.
 * [REST]: Updated http status code returned (400 Bad Request) when uploading files to the wrong type of run
 
 0.22.0 to 19.01